Validate that AutoCheck is prepended
