+++ lib/msf/core/exploit/auto_check.rb
@@ -3,6 +3,10 @@
 module Msf
 module Exploit::Remote::AutoCheck
 
+  def self.included(base)
+    raise NotImplementedError, "#{self.name} should not be included, it should be prepended"
+  end
+
   def initialize(info = {})
     super
