+++ modules/exploits/windows/local/cve_2019_1458_wizardopium.rb
@@ -15,7 +15,7 @@ class MetasploitModule < Msf::Exploit::Local
   include Msf::Post::Windows::Priv
   include Msf::Post::Windows::FileInfo
   include Msf::Post::Windows::ReflectiveDLLInjection
-  include Msf::Exploit::Remote::AutoCheck
+  prepend Msf::Exploit::Remote::AutoCheck
 
   def initialize(info = {})
     super(
@@ -122,8 +122,6 @@ class MetasploitModule < Msf::Exploit::Local
   end
 
   def exploit
-    super
-
     if is_system?
       fail_with(Failure::None, 'Session is already elevated')
     end
