+++ modules/exploits/windows/nimsoft/nimcontroller_bof.rb
@@ -7,7 +7,7 @@ class MetasploitModule < Msf::Exploit::Remote
   Rank = ExcellentRanking
 
   include Msf::Exploit::Remote::Tcp
-  include Msf::Exploit::Remote::AutoCheck
+  prepend Msf::Exploit::Remote::AutoCheck
 
   def initialize(info = {})
     super(
@@ -105,8 +105,6 @@ class MetasploitModule < Msf::Exploit::Remote
   end
 
   def exploit
-
-    super
     connect
 
     shellcode = make_nops(500)
